Draft: change argument to angle in make_radial_dimension
Instead of the fourth argument being a point through which
the dimension line goes through, the argument is the angle of
apperture of the dimension line, which is a radius
of the measured arc edge.

This currently doesn't work because the angle changes depending
on the current working plane. This angle needs to be calculated
taking into account the current working plane, which includes
adjusting the cosine and sine of the normal of the plane.